+++ src/features/enrollment/hooks/index.ts
@@ -3,3 +3,4 @@ export * from "./use-enrollment-type-switch";
 export * from "./use-courses-query";
 export * from "./use-enrollment-draft";
 export * from "./use-enrollment-mutation";
+export * from "./use-prevent-unsaved-changes";